Moved the API key and Streamer List view into separate components. Parent App component now passes state and functions to children as props.
